@@ -9,30 +9,29 @@
         "lint": "next lint"
     },
     "dependencies": {
-        "@heroicons/react": "^2.0.18",
-        "@portabletext/react": "^3.0.9",
-        "@sanity/image-url": "^1.0.2",
-        "@sanity/vision": "^3.17.0",
-        "eslint": "8.37.0",
-        "eslint-config-next": "^14.1.3",
+        "@heroicons/react": "^2.2.0",
+        "@portabletext/react": "^5.0.0",
+        "@sanity/image-url": "^2.0.0",
+        "@sanity/vision": "^4.19.0",
         "googleapis": "^166.0.0",
-        "next": "^14.1.3",
+        "next": "16.0.5",
         "next-auth": "^4.24.7",
-        "next-sanity": "^8.3.2",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "sanity": "^3.32.0",
-        "styled-components": "^5.3.11",
-        "typescript": "5.2.2"
+        "next-sanity": "^11.6.10",
+        "react": "19.2.0",
+        "react-dom": "19.2.0",
+        "sanity": "^4.19.0",
+        "styled-components": "^6.1.19",
+        "typescript": "^5.9.3"
     },
     "devDependencies": {
-        "@sanity/eslint-config-studio": "^2.0.4",
-        "@types/node": "20.8.3",
-        "@types/react": "18.2.25",
-        "@types/react-dom": "18.2.11",
-        "@types/styled-components": "^5.1.28",
-        "autoprefixer": "^10.4.16",
-        "postcss": "^8.4.21",
-        "tailwindcss": "^3.3.3"
+        "@sanity/eslint-config-studio": "^5.0.2",
+        "@tailwindcss/postcss": "^4.1.17",
+        "@types/node": "^24.10.1",
+        "@types/react": "19.2.7",
+        "@types/react-dom": "19.2.3",
+        "@types/styled-components": "^5.1.36",
+        "eslint": "^9.39.1",
+        "eslint-config-next": "^16.0.5",
+        "postcss": "^8.5.6"
     }
 }
