@@ -1,9 +1,12 @@
 import { client } from "./sanity.client";
-import imageUrlBuilder from "@sanity/image-url";
+import {
+    createImageUrlBuilder,
+    type SanityImageSource,
+} from "@sanity/image-url";
 
-const builder = imageUrlBuilder(client);
+const builder = createImageUrlBuilder(client);
 
-function urlFor(source: any) {
+export function urlFor(source: SanityImageSource) {
     return builder.image(source);
 }
 
